refactor: move word type
