+++ src/features/game/card/WordCard.ts
@@ -6,11 +6,21 @@ import RoundState, { type MoveCardAction } from "../model/RoundState";
 import HintSettings from "../model/HintSettings";
 
 import type { RowData } from "../fields/Row";
-import { Word, RowType } from "../types";
+import { RowType } from "../types";
 import { assertNonNull, IMAGES_BASE_URL } from "../../../shared/helpers";
 
 import styles from "./WordCard.module.css";
 import rowStyles from "../fields/Row.module.css";
+
+export interface Word {
+  readonly correctPosition: number;
+  text: string;
+  width: number;
+  isLast: boolean;
+  offset: number;
+  stage: number;
+  image: string;
+}
 
 export default class WordCard extends Draggable {
   private textSpan: Component;
