+++ src/features/game/model/RoundState.ts
@@ -2,7 +2,8 @@ import State from "../../../app/state/StatePublisher";
 import RoundSettings, { RoundResult } from "./RoundSettings";
 import { Observer, Publisher } from "../../../shared/Observer";
 
-import { RowType, StageStatus, Word } from "../types";
+import { RowType, StageStatus } from "../types";
+import { type Word } from "../card/WordCard";
 import {
   generateStageWords,
   prepareRound,
