+++ src/features/game/types.ts
@@ -9,13 +9,3 @@ export enum StageStatus {
   INCORRECT = "incorrect",
   AUTOCOMPLETED = "autocompleted",
 }
-
-export interface Word {
-  readonly correctPosition: number;
-  text: string;
-  width: number;
-  isLast: boolean;
-  offset: number;
-  stage: number;
-  image: string;
-}
